requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==4.2.0
django-ninja==0.22.1
pydantic==1.10.8
sqlparse==0.4.3
typing_extensions==4.6.3
